requirements.txt
# 2026 年最新角膜地形图分类模型 - 依赖包
# PyTorch 核心（最新稳定版）
torch>=2.4.0
torchvision>=0.19.0

# 模型库（包含 2025-2026 最新模型）
timm>=1.0.0

# 数据处理
opencv-python>=4.10.0
pillow>=10.0.0
scikit-learn>=1.3.0

# 可视化和进度条
matplotlib>=3.7.0
tqdm>=4.65.0

# 工具
numpy>=1.24.0
pandas>=2.0.0

# Web 部署（新增）
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.0
